+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `lora` and `fsk` cargo features
 - `set_power_amplification` method
 - `set_pa_ramp` method
+- `set_ocp_trim`, `turn_ocp_off` and `turn_ocp_on` methods
 
 ## [0.1.0] - 2026-03-07